Moved game api apart
