+++ src/api/app.ts
@@ -1,9 +1,9 @@
 import express from "express"
 import cookieParser from "cookie-parser"
 import cors from 'cors'
-import wizard from "../api/games/wizard"
-import player from "../api/player"
-import room from "../api/room"
+import wizard from "../wizard_game/api/wizard"
+import player from "./player"
+import room from "./room"
 import httpLogger from "morgan"
 import logger from "../engine/winston"
 import assertPlayer from "../engine/request_handlers/player_assert"
